Fix two abilities fields being u8s

+++ src/battle_controllers.c
@@ -872,7 +872,7 @@ void BtlController_EmitChooseItem(u8 bufferId, u8 *battlePartyOrder)
     PrepareBufferDataTransfer(bufferId, sBattleBuffersTransferData, 4);
 }
 
-void BtlController_EmitChoosePokemon(u8 bufferId, u8 caseId, u8 slotId, u8 abilityId, u8 *data)
+void BtlController_EmitChoosePokemon(u8 bufferId, u8 caseId, u8 slotId, u16 abilityId, u8 *data)
 {
     s32 i;
